requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6